Add command line parser
Prior to this, we took input from in.md (because it was easy for a
prototype) and wrote output to stdout (similarly, because it was easy,
but also because I think that's the right default.

This changeset adds a required argument for the input file and an
optional option for output. Output still defaults to stdout.