Fix bug with empty attributes
Attributes should only be empty when we have empty input, but that
doesn't mean we shouldn't handle the case correctly. I noticed this
when I accidentally input an empty file and got an error. Since `fodlr1`
doesn't use the `mempty` instance, but instead just assumes that there's
at least one in the `Foldable`, it fatally errored.

I think there's a verison that takes advantage of `Monoid`, but it was
easier to just fill it in with the empty set here.