Refactor parsing for modularity with State monad
The parsing algorithm was not written in a particularly modular way: the
`parseSummary` control flow went directly into the `parseDescription`
function, tightly coupling the two. The parsing algorithm was already
written in a way that perfectly modeled the State monad, so when I
refactored for modularity, I also refactored to actually use the State
monad. This conveniently led to the `parseAll` function to just be an
iteration until the remaining Pandoc blocks list was empty.

Refactoring to use State made it trivial to parse the summary and the
body in two separate steps. This should hopefully make for an easier
code read.